# Avoid hashing the operator type in the GraphViewer priority node check unless the string has a chance of matching.