--- a/onnxruntime/core/graph/graph_viewer.cc
+++ b/onnxruntime/core/graph/graph_viewer.cc
@@ -18,11 +18,14 @@ bool NodeCompare::operator()(const Node* n1, const Node* n2) const {
 #if !defined(ORT_MINIMAL_BUILD)
 struct PriorityNodeCompare {
   inline bool IsHighPri(const Node* n) const {
-    static const std::unordered_set<std::string> high_pri_ops = {"Shape", "Size"};
     const auto& op_type = n->OpType();
     const auto op_type_len = op_type.size();
-    // string hashing for the 'find' is expensive, so avoid if the op length isn't 4 or 5.
-    return (op_type_len < 6 && op_type_len > 3) ? high_pri_ops.find(n->OpType()) != high_pri_ops.end() : false;
+
+    // local statics so we can compare std::strings in the checks
+    static const std::string shape_op("Shape");
+    static const std::string size_op("Size");
+
+    return op_type == shape_op || op_type == size_op;
   }
 
   // Used for std::priority_queue